Switch to postgres

+++ entity/src/entities/hacker.rs
@@ -9,7 +9,7 @@ pub struct Model {
     #[sea_orm(primary_key, auto_increment = false)]
     pub discord_id: String,
     pub username: String,
-    pub fk_team_id: Option<i32>,
+    pub fk_team_id: Option<i64>,
 }
 
 #[derive(Copy, Clone, Debug, EnumIter, DeriveRelation)]
